final_synthetic_data/requirements.txt
blenderproc
numpy<2.0
Pillow
anomalib[full]
ultralytics
pyyaml 